Move FunctionsCodingIP to separate file
